requirements.txt
SpeechRecognition==3.10.0
pyttsx3==2.90
GoogleBard==2.1.0
PyAudio==0.2.14
